Add files via upload
This is what I put together for my final project for my Advanced Java Programming class. The Assignment was to make an application using Spring Boot for a campus testing center allowing students to sign in. I made  a web interface for students to sign in by providing their first name, last name, email, and selecting a course from a dropdown menu. The backend uses an H2 in-memory database to store this information, with the Spring Data JPA handling database interactions. The application uses Thymeleaf as the template engine to render the web pages and connect the data between the web interface and the database. In the video, I opened the H2 console to test that the sign-in information was being stored as intended. It wasn't part of the assignment to style the web form, but I couldn't help myself, CSS is fun.